float to np.float32

+++ vedo/dolfin.py
@@ -670,7 +670,7 @@ class MeshActor(Mesh):
         movedpts = coords + deltas
         if movedpts.shape[1] == 2:  # 2d
             movedpts = np.c_[movedpts, np.zeros(movedpts.shape[0])]
-        self.polydata(False).GetPoints().SetData(utils.numpy2vtk(movedpts, dtype=float))
+        self.polydata(False).GetPoints().SetData(utils.numpy2vtk(movedpts, dtype=np.float32))
         self.polydata(False).GetPoints().Modified()
